add support for persistence through DETS add persistence tests
when launched with persistent: true the ExRated server
will persist data on disk in a file with the same name
of ets_table_name.
When started with persistence activated, ExRated will reload
the data from the disk, store it in memory and continue from there.
When stopped (actually when terminate is executed), the data will
be persisted back on disk.